Started adding tests for xml parsing
