+++ scripts/idl/BUILD.gn
@@ -71,5 +71,6 @@ pw_python_package("idl") {
   tests = [
     "test_matter_idl_parser.py",
     "test_generators.py",
+    "test_xml_parser.py",
   ]
 }
